refactor(frontend): move all server stuff into one file
