+++ apps/frontend/app/components/toaster.tsx
@@ -1,7 +1,7 @@
 import { Notifications, notifications } from "@mantine/notifications";
 import { useEffect } from "react";
 import { match } from "ts-pattern";
-import type { Toast } from "~/lib/toast.server";
+import type { Toast } from "~/lib/utilities.server";
 
 export function Toaster({ toast }: { toast?: Toast | null }) {
 	return (
