+++ apps/frontend/app/routes/_dashboard.collections.list.public.tsx
@@ -23,9 +23,8 @@ import {
 	ApplicationPagination,
 	DebouncedSearchInput,
 } from "~/components/common";
-import { gqlClient } from "~/lib/api.server";
 import { useGetMantineColor, useSearchParam } from "~/lib/hooks";
-import { getCoreDetails } from "~/lib/utilities.server";
+import { getCoreDetails, gqlClient } from "~/lib/utilities.server";
 
 const searchParamsSchema = z.object({
 	page: zx.IntAsString.optional(),
