+++ apps/frontend/app/routes/_dashboard.fitness.measurements.list.tsx
@@ -47,12 +47,16 @@ import { namedAction } from "remix-utils/named-action";
 import { match } from "ts-pattern";
 import { z } from "zod";
 import { zx } from "zodix";
-import { getAuthorizationHeader, gqlClient } from "~/lib/api.server";
 import events from "~/lib/events";
 import { ApplicationKey, dayjsLib, redirectToQueryParam } from "~/lib/generals";
 import { useSearchParam } from "~/lib/hooks";
-import { createToastHeaders } from "~/lib/toast.server";
-import { getUserPreferences, processSubmission } from "~/lib/utilities.server";
+import {
+	createToastHeaders,
+	getAuthorizationHeader,
+	getUserPreferences,
+	gqlClient,
+	processSubmission,
+} from "~/lib/utilities.server";
 
 enum TimeSpan {
 	Last7Days = "Last 7 days",
