+++ apps/frontend/app/routes/_dashboard.fitness.workouts.$id._index.tsx
@@ -67,11 +67,16 @@ import { match } from "ts-pattern";
 import { withFragment } from "ufo";
 import { z } from "zod";
 import { DisplayExerciseStats } from "~/components/fitness";
-import { getAuthorizationHeader, gqlClient } from "~/lib/api.server";
 import { dayjsLib, getSetColor } from "~/lib/generals";
 import { getWorkoutStarter, useGetMantineColor } from "~/lib/hooks";
-import { createToastHeaders, redirectWithToast } from "~/lib/toast.server";
-import { getUserPreferences, processSubmission } from "~/lib/utilities.server";
+import {
+	createToastHeaders,
+	getAuthorizationHeader,
+	getUserPreferences,
+	gqlClient,
+	processSubmission,
+	redirectWithToast,
+} from "~/lib/utilities.server";
 import { duplicateOldWorkout, getExerciseDetails } from "~/lib/workout";
 
 export const loader = async ({ request, params }: LoaderFunctionArgs) => {
