+++ apps/frontend/app/routes/_dashboard.media.genre.$id._index.tsx
@@ -13,9 +13,8 @@ import { z } from "zod";
 import { zx } from "zodix";
 import { ApplicationGrid, ApplicationPagination } from "~/components/common";
 import { BaseDisplayItem } from "~/components/media";
-import { gqlClient } from "~/lib/api.server";
 import { useSearchParam } from "~/lib/hooks";
-import { getCoreDetails } from "~/lib/utilities.server";
+import { getCoreDetails, gqlClient } from "~/lib/utilities.server";
 
 const searchParamsSchema = z.object({
 	page: zx.IntAsString.default("1"),
