+++ apps/frontend/app/routes/_dashboard.media.genre.list.tsx
@@ -25,9 +25,8 @@ import {
 	ApplicationPagination,
 	DebouncedSearchInput,
 } from "~/components/common";
-import { gqlClient } from "~/lib/api.server";
 import { useGetMantineColor, useSearchParam } from "~/lib/hooks";
-import { getCoreDetails } from "~/lib/utilities.server";
+import { getCoreDetails, gqlClient } from "~/lib/utilities.server";
 
 const searchParamsSchema = z.object({
 	page: zx.IntAsString.default("1"),
