+++ apps/frontend/app/routes/_dashboard.media.groups.list.tsx
@@ -24,9 +24,8 @@ import {
 	DebouncedSearchInput,
 } from "~/components/common";
 import { BaseDisplayItem } from "~/components/media";
-import { gqlClient } from "~/lib/api.server";
 import { useSearchParam } from "~/lib/hooks";
-import { getCoreDetails } from "~/lib/utilities.server";
+import { getCoreDetails, gqlClient } from "~/lib/utilities.server";
 
 const searchParamsSchema = z.object({
 	page: zx.IntAsString.default("1"),
