+++ apps/frontend/app/routes/_dashboard.settings.miscellaneous.tsx
@@ -20,9 +20,13 @@ import {
 	UserLot,
 } from "@ryot/generated/graphql/backend/graphql";
 import { z } from "zod";
-import { getAuthorizationHeader, gqlClient } from "~/lib/api.server";
-import { createToastHeaders } from "~/lib/toast.server";
-import { getUserDetails, processSubmission } from "~/lib/utilities.server";
+import {
+	createToastHeaders,
+	getAuthorizationHeader,
+	getUserDetails,
+	gqlClient,
+	processSubmission,
+} from "~/lib/utilities.server";
 
 export const loader = async ({ request }: LoaderFunctionArgs) => {
 	const [userDetails] = await Promise.all([getUserDetails(request)]);
