+++ apps/frontend/app/routes/_dashboard.settings.notifications.tsx
@@ -39,10 +39,14 @@ import { match } from "ts-pattern";
 import { z } from "zod";
 import { zx } from "zodix";
 import { confirmWrapper } from "~/components/confirmation";
-import { getAuthorizationHeader, gqlClient } from "~/lib/api.server";
 import { dayjsLib } from "~/lib/generals";
-import { createToastHeaders } from "~/lib/toast.server";
-import { getCoreDetails, processSubmission } from "~/lib/utilities.server";
+import {
+	createToastHeaders,
+	getAuthorizationHeader,
+	getCoreDetails,
+	gqlClient,
+	processSubmission,
+} from "~/lib/utilities.server";
 
 export const loader = async ({ request }: LoaderFunctionArgs) => {
 	const [coreDetails, { userNotificationPlatforms }] = await Promise.all([
