+++ apps/frontend/app/routes/_dashboard.settings.preferences.tsx
@@ -52,9 +52,13 @@ import { match } from "ts-pattern";
 import { withQuery, withoutHost } from "ufo";
 import { z } from "zod";
 import { zx } from "zodix";
-import { getAuthorizationHeader, gqlClient } from "~/lib/api.server";
 import { redirectToQueryParam } from "~/lib/generals";
-import { getUserDetails, getUserPreferences } from "~/lib/utilities.server";
+import {
+	getAuthorizationHeader,
+	getUserDetails,
+	getUserPreferences,
+	gqlClient,
+} from "~/lib/utilities.server";
 import classes from "~/styles/preferences.module.css";
 
 const searchSchema = z.object({
