+++ apps/frontend/app/routes/_dashboard.tsx
@@ -49,8 +49,8 @@ import {
 import { produce } from "immer";
 import { joinURL } from "ufo";
 import { HiddenLocationInput } from "~/components/common";
-import { redirectIfNotAuthenticated } from "~/lib/api.server";
 import { ApplicationKey, getLot } from "~/lib/generals";
+import { redirectIfNotAuthenticated } from "~/lib/utilities.server";
 import {
 	colorSchemeCookie,
 	expectedEnvironmentVariables,
