+++ apps/frontend/app/routes/auth.register.tsx
@@ -15,9 +15,13 @@ import {
 } from "@ryot/generated/graphql/backend/graphql";
 import { match } from "ts-pattern";
 import { z } from "zod";
-import { getIsAuthenticated, gqlClient } from "~/lib/api.server";
-import { createToastHeaders, redirectWithToast } from "~/lib/toast.server";
-import { getCoreEnabledFeatures } from "~/lib/utilities.server";
+import {
+	createToastHeaders,
+	getCoreEnabledFeatures,
+	getIsAuthenticated,
+	gqlClient,
+	redirectWithToast,
+} from "~/lib/utilities.server";
 import classes from "~/styles/auth.module.css";
 
 export const loader = async ({ request }: LoaderFunctionArgs) => {
